[MDB Ignore] Job Description Audit; Job Audit... Really, Just An Audit (#19)
* Adventurer Description Audit

* church awesome

* Courtier

* garrison

* Nobles

* Kills nightmaidens and nightmen. With hammers.

Same reason we removed consort and prince, sexo and any implicit relationships should be safe, opt-in, and not mechanically incentivized in ANY fashion. This is all of the above, grimshart; etc et al. The bathhouse has been subsequentially massive reduced in scale and made a public space. I'll probably clean it up more later; but I rather focus on the NEW map verse this one anyways

* oop

* all but prisoner jobs

* oop

* ALL JOBS DONE