Reverts Pilgrim Round Start and Adds Outlander Virtue (#51)
* Renable Roundstart Pilgrim + Adds Outlander Virtue

Allows for Pilgrim Roundstart.

Additionally, adds the Outlander virtue. This virtue marks you as a foreigner, gives you a small pouch of rations, and gives you 2 levels in traversal skills from your long trek to The Ridge. These skills cannot exceed 4, and so are of little value to most combat roles. Command Roles are incompatible with the Outlander Virtue.

* nudges the skills further, as 4 climbing is too much

nudges the skills further, as 4 climbing is too much

* Update priest.dm

---------

Co-authored-by: BluBerry016 <50649185+unit0016@users.noreply.github.com>