[QoL] Adjust Magic changes based on the Feedback Requests (#137)
* Initial Commit

Makes a new helper to check for combat class traits being innate, as spells now give you the ability to get some of these features, I think its better to avoid crazy scenarios and unintuitive mechanics. (Plus it looks cleaner)

* another helper

* Re-adds fatigue based XP to casting spells

We doubled the original multiplier and made it so only you can level it if you have a level already in it.

* Add divine helper

* Adds support for Divine and Arcane Crafting

* Makes SCOM need arcane support, uses crafting instead of Arcane skill.

* Tentative Proposal for Arcane Classes

* removes duplicate smelting to janitor

* Proposal Feedback

* Special Reward on Legendary Arcane

* Adds Arcane tutoring

* Adds adjustment to death knight that was missed

* Magic staffs drop gems when destroyed

* Adventurer Clerics fix and Starting Devotion Fix

* Sets Alchemist positions to 0 until its removed or fixed